requirements.txt
flask
onnxruntime
pillow
numpy
dvc